Solution
According to reports, the water usage rate of Saudi Arabia is twice the world average. To deal with this alarming situation, the Department of Statistics has collected the total water usage data from various cities in different regions for the recent years along with the population details during these years. You are required to develop an application that can use this data as input and provide insight details of the water usage from different perspective such as the per year percentage increase or decrease in water usage over time in a specific city, the per capita water usage, the lowest and highest water usage cities etc. All this information will be helpful to help in reducing the water usage per capita.